# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"biblioteca-steam","description":"biblioteca-steam application","contact":{"name":"Agustin Nieto","email":"santiago-n@live.com.ar"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"biblioteca-steam application"}],"paths":{"/usuarios/login":{"post":{"tags":["usuario-controller"],"summary":"Logea un usuario a la plataforma","description":"Devuelve el usuario logeado si sus credenciales son correctas.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credentials"}}},"required":true},"responses":{"404":{"description":"Email and Password are not in the database!","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usuario"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usuario"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Usuario"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app is alive will respond with pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/juegos/{nombre}/":{"get":{"tags":["juego-controller"],"operationId":"getJuegosByNombre","parameters":[{"name":"nombre","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Juego"}}}}}}},"/juegos/genero/{genero}/":{"get":{"tags":["juego-controller"],"operationId":"getJuegosByGenero","parameters":[{"name":"genero","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Juego"}}}}}}}},"/bibliotecas/{horas}/":{"get":{"tags":["biblioteca-controller"],"operationId":"findByHoras","parameters":[{"name":"horas","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Biblioteca"}}}}}}}}},"components":{"schemas":{"Credentials":{"required":["password"],"type":"object","properties":{"email":{"title":"User's email","type":"string","description":"Your account email","example":"example@mail.com"},"password":{"type":"string"}}},"Usuario":{"required":["nombre"],"type":"object","properties":{"id":{"title":"Usuario ID","type":"integer","description":"El identificador del Usuario","format":"int64","example":7},"email":{"type":"string"},"password":{"type":"string"},"nombre":{"title":"Nombre del Usuario","type":"string","description":"El nombre del usuario","example":"MiNombre"},"pais":{"type":"string"},"lastLogin":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"nivel":{"type":"integer","format":"int32"}}},"Genero":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}},"Juego":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"releaseDate":{"type":"string","format":"date-time"},"precio":{"type":"integer","format":"int32"},"genero":{"$ref":"#/components/schemas/Genero"},"logros":{"type":"integer","format":"int32"},"developer":{"type":"string"},"publisher":{"type":"string"},"rating":{"type":"string"}}},"Biblioteca":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"juego":{"$ref":"#/components/schemas/Juego"},"usuario":{"$ref":"#/components/schemas/Usuario"},"horas":{"type":"integer","format":"int32"},"logros":{"type":"integer","format":"int32"},"review":{"type":"string"}}}}}}