# src/main/java/tup/bibliotecasteam/models/Genero.java
package tup.bibliotecasteam.models;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class Genero {
    private Long id;
    private String nombre;
}